requirements.txt
pandas==1.2.4
numpy==1.22.3
scikit-image==0.18.1
TPOT==0.11.7
gunicorn==20.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
Werkzeug==0.16.0
matplotlib==3.1.3
seaborn==0.9.0
Flask-SQLAlchemy==2.5.1


